Package Orckestra.Media.StaticFileLinkMediaProvider:
Core project for new package Orckestra.Media.StaticFileLinkMediaProvider - this is a (read only) media provider that show up in the media structure, exposing files from website and allowing console users to link directly to static files. Implements "IMediaUrlProvider" to ensure links are mapped to normal relative paths (and as such this provider is safe to point even to the root of the website - it will not serve any protected files, since it does not serve files, just paths).

To register the assembly raw, copy it to ~/bin and add this line inside COmposite.config <DataProviderPlugins /> element:

<add basePath="~/download" storeTitle="Static files from ~/download" storeDescription="" storeId="download" name="StaticFileLinkMediaProvider" type="Orckestra.Media.StaticFileLinkMediaProvider.MediaProvider, Orckestra.Media.StaticFileLinkMediaProvider" />