Uploading a tool to change a structure of typescripts
Connected with task #27664. Purpose of this tool: to move typescripts from one location into another one and to regenerate relative references inside files according to the new locations. A background of this task is to move typescripts of RefApp into new general location. Afterwards here will be also added a C1 package for dynamic typescripts compilation.